backend/retranscription/requirements.txt
# Dépendances minimales pour la transcription audio
google-cloud-speech==2.27.0
pyaudio==0.2.14
python-dotenv==1.0.1
